PHOENIX-4259 Split up IndexExtendedIT
